chatgpt_server/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.6
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
exceptiongroup==1.1.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itypes==1.2.0
Jinja2==3.1.2
markdownify==0.11.6
MarkupSafe==2.1.2
multidict==6.0.4
oauthlib==3.2.2
openai==0.26.5
outcome==1.2.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.7.1
PyVirtualDisplay==3.0
requests==2.28.2
requests-oauthlib==1.3.1
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.4.6
uritemplate==4.1.1
urllib3==1.26.14
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
zipp==3.13.0
